refactor: mirror tests/llm/ onto app/llm/'s subpackage structure
