+++ backend/tests/llm/intake/test_exceptions.py
@@ -2,18 +2,10 @@ import pytest
 from fastapi import HTTPException
 
 from app.llm.intake.exceptions import raise_hf_opening_response_missing_text
-from app.llm.outreach.exceptions import raise_outreach_email_empty
 
 
 class TestLlmIntakeExceptions:
     def test_opening_response_missing_text_raises_502(self):
         with pytest.raises(HTTPException) as info:
             raise_hf_opening_response_missing_text()
-        assert info.value.status_code == 502
-
-
-class TestLlmOutreachExceptions:
-    def test_outreach_email_empty_raises_502(self):
-        with pytest.raises(HTTPException) as info:
-            raise_outreach_email_empty()
         assert info.value.status_code == 502
