+++ backend/tests/llm/providers/test_exceptions.py
